Struct Hole: 8 bytes hole in class
1 - Reorder the class members to remove a struct
 hole. this was discovered with poke-a-hole, a really
 nice tool to verify flaws in classes and structs.

2 - Remove a const QList from the class. Probably this
will not affect much, but it's a const QList, that means that
if two instances of this class exists at a time, double of
the memory of the QList will be loaded. A good rule of thumb
is to move every single const with default values (a const that's
being initialized with a constant instead of a bariable) or static
values from the class to the .cpp file. This gives us two things:
2.1 - Less space per object is used
2.2 - The compilation of the software is faster as there's no need
to recompile when there's a change in those variables, it will only
affect it's own compilation unit.