editorconfig: Add initial version
EditorConfig is a special file that's handled by many editors and
can keep the configuration of the specifics of the project out of
the editor settings. This is really userfull if you work in more
than one project at a time, like we do.

KDevelop, Qt Creator, Vim, Emacs, Visual Studio Code, all have
native or plugin support for editorconfig.