seriallink: Add setBaudrate function and ABR detection procedure
setBaudrate function is necessary to avoid developers to forget the changes
in the link configuration, this function provides for us the possibility to change
the baud rate of the serial port and also updates the linkConfiguration for us,
allowing notifications that the configuration has changed.
It also makes the ABR (Auto Baud Rate detection) more modular with a function.

Signed-off-by: Patrick José Pereira <patrickelectric@gmail.com>