CMake: use full patch for includes unless it's in the same folder
Add every include in the same "namespace" is misleading, and makes
it worse to reason about how the software is structured. In cmake
I added only the toplevel folder as include path (the src folder)

In the .pro every single directory is added as include path, so you
can easily use an include of one subfoler in another without realizing
it's in a different folder and probably should not be used there.