Ping360Visualizer: Makes vertical and horizontal flip mutually exclusive for 360 sector
It's not possible to accomplish the same thing with ButtonGroup,
since there functionality depends of 3 variables changing how vertical flip works.
There is also the problem that when the button is not visible, the enabled component
of such control is also false, so it makes the onEnabled logic harder.
This approach uses the binding method to accomplish the same behavior.

Signed-off-by: Patrick José Pereira <patrickelectric@gmail.com>